working-version/requirements.txt
# NOTE: torch/torchvision are installed via scripts/setup/install_dependencies.sh
# to ensure the wheel matches the installed CUDA toolkit (default: PyTorch 2.9.0+cu130)
# TensorRT-LLM 1.2.0rc5 requires CUDA 13.0 and torch 2.9.0

# NOTE: TensorRT-LLM (installed separately via wheel)
# tensorrt-llm==1.2.0rc5

# Core server dependencies - versions must match TRT-LLM 1.2.0rc5 requirements
fastapi==0.121.3
starlette==0.49.1
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
websockets==12.0

# HuggingFace and ML - versions MUST match TRT-LLM 1.2.0rc5 requirements exactly
huggingface_hub==0.36.0
transformers==4.56.0
numpy==1.26.4
snac==1.2.1
triton==3.5.0

# Audio processing
soundfile==0.12.1

# TensorRT-LLM dependencies
cuda-python==13.0.0
hf_transfer==0.1.8

# Build tools - setuptools MUST be <80 for TRT-LLM 1.2.0rc5
ninja==1.13.0
packaging==25.0
setuptools==79.0.0
wheel==0.45.1